mitsubishi_test/requirements.txt
gradio
psycopg2-binary
pandas
openpyxl
requests
pymupdf
pillow
rapidfuzz
